Filter the single page-open event before turning off analytics
If dashboard is opened and then the analytics disable parameter is added, the page is refreshed, meaning we get 2 page opens. Second page open with the toggle is filtered but first isn't. This commit includes deleting the first one too.
I am aware this could result in data inconsistencies because this is not a solid way of doing it, and a fix is on the todo-list. For now this will be okay because the chance of producing a data inconsistency is relatively small.